# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZeeoWork API Name",
        "description": "ZeeoWork API Description",
        "contact": {
            "email": "tech@zeework.ir"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://devpandora.zeework.ir/"
        }
    ],
    "paths": {
...
        "/api/hr/profile": {
            "get": {
                "tags": [
                    "OrganizationAdmin/Profile"
                ],
                "summary": "Get current OrganizationAdmin profile",
                "operationId": "9223cc8c5c6aa75e8717651372088c0b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HrOrganizationAdminResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "OrganizationAdmin/Profile"
                ],
                "summary": "Update current OrganizationAdmin profile",
                "operationId": "5b5300584b4aebd43585e7a6f2885311",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrganizationAdminProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "خروجی استاندارد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/hr/profile/change-password": {
            "put": {
                "tags": [
                    "OrganizationAdmin/Profile"
                ],
                "summary": "Change current OrganizationAdmin password",
                "operationId": "fd9bd90f6eee2736508834a5e6b2bc2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeOrganizationAdminPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "خروجی استاندارد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/physician/profile": {
            "get": {
                "tags": [
                    "Physician/Profile"
                ],
                "summary": "Get current Physician profile",
                "operationId": "9f8b3964b308e1a7bdbf882d6ea3add0",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhysicianSelfResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Physician/Profile"
                ],
                "summary": "Update current Physician profile",
                "operationId": "64849a8f5014a537aedd45336a6bfaa6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePhysiciansProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "خروجی استاندارد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/physician/profile/change-password": {
            "put": {
                "tags": [
                    "Physician/Profile"
                ],
                "summary": "Change current Physician password",
                "operationId": "edad21f88990dba26368424da86712ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePhysicianPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "خروجی استاندارد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": true
                    },
                    "messages": {
                        "description": "آرایه‌ای از پیام‌ها برای کاربر (می‌تواند شامل پیام‌های موفقیت، خطا یا اطلاعاتی باشد)",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "عملیات با موفقیت انجام شد."
                        },
                        "example": [
                            "عملیات موفقیت‌آمیز بود",
                            "پیام دوم در صورت وجود"
                        ],
                        "nullable": true
                    },
                    "data": {
                        "type": "object",
                        "example": {
                            "id": 1,
                            "name": "Ali"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ChangeOrganizationAdminPasswordRequest": {
                "title": "Change OrganizationAdmin Password Request",
                "required": [
                    "current_password",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "current_password": {
                        "description": "Admin's current password",
                        "type": "string",
                        "format": "password",
                        "example": "current_secret"
                    },
                    "password": {
                        "description": "Admin's new password",
                        "type": "string",
                        "format": "password",
                        "example": "NewS3cr3t!"
                    },
                    "password_confirmation": {
                        "description": "Confirmation of the new password",
                        "type": "string",
                        "format": "password",
                        "example": "NewS3cr3t!"
                    }
                },
                "type": "object"
            },
...
            "UpdateOrganizationAdminProfileRequest": {
                "title": "Update OrganizationAdmin Profile Request",
                "required": [],
                "properties": {
                    "name": {
                        "description": "Admin's first name",
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "description": "Admin's email",
                        "type": "string",
                        "example": "Doe@test.com"
                    },
                    "phone": {
                        "description": "Admin's phone",
                        "type": "string",
                        "example": "09012211111"
                    }
                },
                "type": "object"
            },
            "ChangePhysicianPasswordRequest": {
                "title": "Change Physician Password Request",
                "required": [
                    "current_password",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "current_password": {
                        "description": "Admin's current password",
                        "type": "string",
                        "format": "password",
                        "example": "current_secret"
                    },
                    "password": {
                        "description": "Admin's new password",
                        "type": "string",
                        "format": "password",
                        "example": "NewS3cr3t!"
                    },
                    "password_confirmation": {
                        "description": "Confirmation of the new password",
                        "type": "string",
                        "format": "password",
                        "example": "NewS3cr3t!"
                    }
                },
                "type": "object"
            },
            "UpdatePhysiciansProfileRequest": {
                "title": "Update physicians Profile Request",
                "required": [],
                "properties": {
                    "name": {
                        "description": "physicians's first name",
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "description": "physicians's email",
                        "type": "string",
                        "example": "Doe@test.com"
                    },
                    "phone": {
                        "description": "physicians's phone",
                        "type": "string",
                        "example": "09012211111"
                    },
                    "photo": {
                        "description": "physicians's avatar image file (optional)",
                        "type": "file",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
            "HrOrganizationAdminResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 42
                    },
                    "name": {
                        "type": "string",
                        "example": "سارا احمدی"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "sara.ahmadi@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "09121234567"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
            "PhysicianSelfResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "name": {
                        "type": "string",
                        "example": "دکتر محمد قاسمی"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ghasemi@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "09123456789"
                    },
                    "photo": {
                        "type": "string",
                        "format": "url",
                        "example": "https://cdn.example.com/physicians/10.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T12:34:56Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}